- (auth) create client/lib/auth for all auth checking for frontend
